Reorganized the whole async code to avoid memory issues
- Replaced all session.get and session.post with async with structure
to close them automatically, they were not closed at all before.

- Now each file is downloaded using chunks to not store a lot of
data in the memory.

- Methods to get and download the data have been replaced by
one function called download_file, it is cleaner this way.

- Moved get_extension into filename_parser because
creating a method for it was not necessary.

- Added a callback function to the download function in order
to know which courses have been completely downloaded in real time
when the attribute show_message in AmeticeBot is set to True.